+++ .github/workflows/aws-generic-deploy.yml
@@ -45,12 +45,8 @@ jobs:
           name: js-sdk-build-artifacts
 
       - name: S3 Deploy
-        run: aws s3 cp ./dist $AWS_S3_BUCKET --recursive && aws s3 cp ./public $AWS_S3_BUCKET --recursive
-        env:
-          AWS_S3_BUCKET: ${{ secrets.aws-s3-bucket }}
+        run: aws s3 cp ./dist/ s3://${{ secrets.aws-s3-bucket }}/v2/ --recursive && aws s3 cp ./cdn-root s3://${{ secrets.aws-s3-bucket }}/ --recursive
 
       - name: Cloudfront Cache Invalidation
         run: |
-          aws cloudfront create-invalidation --distribution-id $AWS_CLOUDFRONT_DISTRIBUTION_ID --paths "/v1/index.js" "/v2/index.js"
-        env:
-          AWS_CLOUDFRONT_DISTRIBUTION_ID: ${{ secrets.aws-cloudfront-distribution-id }}
+          aws cloudfront create-invalidation --distribution-id ${{ secrets.aws-cloudfront-distribution-id }} --paths "/v1/index.js" "/v2/index.js"
