+++ .gitignore
@@ -3,6 +3,7 @@ e2e-app/evervault.bundle.js
 node_modules
 .idea
 .env
+.env.*
 .DS_STORE
 /test-results/
 /playwright-report/
