--- 404.html
@@ -1,5 +1,0 @@
-{
-    "error": 404,
-    "status": "Page not found",
-    "evervault-problem": "possibly!"
-}
