+++ index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+  <head>
+
+  </head>
+  <body>
+    evervault Test Hello world
+    <div id="ev-inputs" />
+
+    <script type="module" src="scaffold.js"></script>
+  </body>
+</html>
