+++ lib/core/crypto.js
@@ -1,7 +1,6 @@
-/** @format */
+import { Buffer } from 'buffer';
 
-const { Datatypes, errors, cryptoUtils } = require("../../utils");
-const { arrayBufferToBase64 } = require("../../utils/datatypes");
+import { Datatypes, errors, cryptoUtils } from "../utils";
 
 const generateBytes = (byteLength) => {
   return new Promise((resolve, reject) => {
@@ -19,7 +18,7 @@ const generateBytes = (byteLength) => {
   });
 };
 
-module.exports = (config, isDebug) => {
+export default function Crypto(config, isDebug) {
   const _encryptObject = async (
     ecdhTeamKey,
     ecdhPublicKey,
@@ -187,7 +186,7 @@ module.exports = (config, isDebug) => {
     return `ev:${isDebug ? "debug:" : ""}${_evVersionPrefix}${
       datatype !== "string" ? ":" + datatype : ""
     }:${base64RemovePadding(keyIv)}:${base64RemovePadding(
-      arrayBufferToBase64(compressedKey)
+      Datatypes.arrayBufferToBase64(compressedKey)
     )}:${base64RemovePadding(encryptedData)}:$`;
   };
 
@@ -264,3 +263,4 @@ module.exports = (config, isDebug) => {
 
   return { encrypt };
 };
+
