+++ lib/core/forms.js
@@ -1,8 +1,6 @@
-/** @format */
+import * as EventEmitter from "events";
 
-const EventEmitter = require("events");
-
-module.exports = (evervault) => {
+function Forms(evervault) {
   const formEmitter = new EventEmitter();
   return {
     handleForm: async function (form) {
@@ -70,3 +68,5 @@ module.exports = (evervault) => {
     events: formEmitter,
   };
 };
+
+export default Forms;
