+++ lib/core/http.js
@@ -1,8 +1,8 @@
-const { errors } = require("../../utils");
+import { errors } from "../utils";
 
-module.exports = (config, teamId, context) => {
+export default function Http(config, teamId, appId, context) {
   if (!window.fetch)
-    throw errors.InitializationError(
+    throw new errors.InitializationError(
       "Your browser is outdated and does not support window.fetch(). Please upgrade it."
     );
   const request = (method, path, baseUrl, headers = {}, data = undefined) => {
@@ -38,7 +38,7 @@ module.exports = (config, teamId, context) => {
     try {
       const { headers, body } = await get(
         config.keysUrl,
-        `${teamId}?context=${context}`
+        `${teamId}/apps/${appId}?context=${context}`
       );
       return {
         ...body,
