+++ lib/core/inputs.js
@@ -1,9 +1,6 @@
-/** @format */
+import { constructSource, calculateHeight } from "../utils";
 
-const constructSource = require("../../utils/constructSource");
-const calculateHeight = require("../../utils/calculateHeight");
-
-module.exports = (config) => {
+export default function Inputs(config) {
   return {
     generate: function (id, settings) {
       document.getElementById(id).innerHTML = `<iframe src="${constructSource(
@@ -32,14 +29,20 @@ module.exports = (config) => {
             window.addEventListener(
               "message",
               (event) => {
-                if (event.origin !== config.input.inputsUrl) return;
+                if (event.origin !== config.input.inputsOrigin) return;
                 fn(event.data);
               },
               false
             );
           }
         },
+        setLabels: (labels) => {
+          const channel = new MessageChannel();
+          document
+            .getElementById("ev-iframe")
+            .contentWindow.postMessage(labels, "*", [channel.port2]);
+        },
       };
     },
   };
-};
+}
