+++ lib/curves/asn1.js
@@ -4,7 +4,7 @@
 
 // Almost every ASN.1 type that's important for CSR
 // can be represented generically with only a few rules.
-var ASN1 = function ASN1(/*type, hexstrings...*/) {
+function ASN1(/*type, hexstrings...*/) {
   var args = Array.prototype.slice.call(arguments);
   var typ = args.shift();
   var str = args.join("").replace(/\s+/g, "").toLowerCase();
@@ -66,4 +66,4 @@ ASN1.numToHex = function (d) {
   return d;
 };
 
-module.exports = ASN1;
+export default ASN1;
