+++ lib/curves/base.js
@@ -1,4 +1,6 @@
-const ASN1 = require("./asn1");
+import { Buffer } from 'buffer';
+
+import ASN1 from './asn1';
 
 /**
  * Given an EC curve name and its constants, generate a DER encoder for its compressed public keys
@@ -6,7 +8,7 @@ const ASN1 = require("./asn1");
  * @param curveValues
  * @returns Function(decompressedPublicKey): base64EncodedString
  */
-const createCurve = (curveName, curveValues) => {
+export function createCurve(curveValues) {
   const asn1Encoder = buildEncoder(curveValues);
   return (decompressedPublicKey) => {
     return asn1Encoder(
@@ -15,7 +17,7 @@ const createCurve = (curveName, curveValues) => {
   };
 };
 
-const buildEncoder = ({ p, a, b, seed, generator, n, h }) => {
+function buildEncoder({ p, a, b, seed, generator, n, h }) {
   return (decompressedKey) => {
     const hexEncodedKey = ASN1(
       "30",
@@ -60,6 +62,3 @@ const buildEncoder = ({ p, a, b, seed, generator, n, h }) => {
   };
 };
 
-module.exports = {
-  createCurve,
-};
