+++ lib/curves/index.js
@@ -1,3 +1,1 @@
-module.exports = {
-  P256: require("./p256"),
-};
+export * as P256 from "./p256";
