+++ lib/curves/p256.js
@@ -1,4 +1,5 @@
-const { createCurve } = require("./base");
+import { createCurve } from "./base.js";
+
 // https://neuromancer.sk/std/x962/prime256v1
 const P256_CONSTANTS = {
   p: "FF FF FF FF 00 00 00 01 00 00 00 00 00 00 00 00 00 00 00 00 FF FF FF FF FF FF FF FF FF FF FF FF",
@@ -11,6 +12,8 @@ const P256_CONSTANTS = {
   h: "01",
 };
 
-module.exports = {
-  encodePublicKey: createCurve("prime256v1", P256_CONSTANTS),
-};
+/**
+ * Encode Public Key with prime256v1 curve
+ */
+export const encodePublicKey = createCurve(P256_CONSTANTS);
+
