+++ lib/main.js
@@ -1,18 +1,38 @@
-const { Datatypes, errors, extractDomain } = require("../utils");
-const Config = require("./config");
-const { Crypto, Http, Forms, Input } = require("./core");
-const { deriveSharedSecret } = require("../utils/crypto");
+import { Buffer } from 'buffer';
 
-class EvervaultClient {
-  constructor(teamId, customConfig = {}) {
+import { Datatypes, errors, extractDomain } from "./utils";
+import Config from "./config";
+import { Crypto, Http, Forms, Input } from "./core";
+import {
+  deriveSharedSecret,
+  buildCageKeyFromSuppliedPublicKey,
+} from "./utils/crypto";
+
+export default class EvervaultClient {
+  constructor(teamId, appId, customConfig = {}) {
     if (!Datatypes.isString(teamId)) {
       throw new errors.InitializationError("teamId must be a string");
     }
-    if (!customConfig) {
-      customConfig = {};
+    if (!Datatypes.isString(appId)) {
+      throw new errors.InitializationError("appId must be a string");
     }
-    this.config = Config(teamId, customConfig.urls);
-    this.crypto = Crypto(this.config.encryption);
+    if (!customConfig) {
+      customConfig = {
+        isDebugMode: false,
+      };
+    }
+
+    if (customConfig.isDebugMode) {
+      this._debugMode = true;
+    }
+
+    this.config = Config(
+      teamId,
+      appId,
+      customConfig.urls,
+      customConfig.publicKey
+    );
+    this.crypto = Crypto(this.config.encryption, this.isInDebugMode());
 
     if (window.location.origin === this.config.input.inputsUrl) {
       this.context = "inputs";
@@ -20,14 +40,27 @@ class EvervaultClient {
       this.context = "default";
     }
 
-    this.http = Http(this.config.http, this.config.teamId, this.context);
+    this.http = Http(
+      this.config.http,
+      this.config.teamId,
+      this.config.appId,
+      this.context
+    );
     this.forms = Forms(this);
     this.forms.register();
     this.input = Input(this.config);
   }
 
   async loadKeys() {
-    const cageKey = await this.http.getCageKey();
+    if (this._ecdhTeamKey) return;
+
+    const cageKey = this.config.encryption.publicKey
+      ? await buildCageKeyFromSuppliedPublicKey(
+          this.config.encryption.publicKey
+        )
+      : this.isInDebugMode()
+      ? this.config.debugKey
+      : await this.http.getCageKey();
     this.defineHiddenProperty(
       "_ecdhTeamKey",
       Buffer.from(cageKey.ecdhP256Key, "base64")
@@ -43,7 +76,11 @@ class EvervaultClient {
     );
     this.defineHiddenProperty("_ecdhPublicKey", keyPair.publicKey);
 
-    this.defineHiddenProperty("_debugMode", cageKey.isDebugMode);
+    // If config forced debug mode don't overwrite
+    if (!this.isInDebugMode()) {
+      this.defineHiddenProperty("_debugMode", cageKey.isDebugMode);
+    }
+
     // recreate crypto module  with debug state
     this.crypto = Crypto(this.config.encryption, this.isInDebugMode());
 
@@ -62,7 +99,7 @@ class EvervaultClient {
     if (Datatypes.isEmptyString(data)) {
       return data;
     }
-    if (!this.keysLoadingPromise && !Datatypes.isDefined(this._ecdhTeamKey)) {
+    if (!this.keysLoadingPromise) {
       this.keysLoadingPromise = this.loadKeys();
     }
     await this.keysLoadingPromise;
@@ -132,4 +169,3 @@ class EvervaultClient {
   }
 }
 
-module.exports = EvervaultClient;
