+++ lib/utils/calculateHeight.js
@@ -1,4 +1,4 @@
-module.exports = function (settings) {
+export default function calculateHeight(settings) {
   if (settings) {
     if (settings.height) {
       return settings.height;
