+++ lib/utils/constructSource.js
@@ -1,4 +1,4 @@
-module.exports = function (config, settings) {
+export default function constructSource(config, settings) {
   let settingsQuery = "";
   if (settings) {
     Object.keys(settings).map((key) => {
@@ -11,4 +11,4 @@ module.exports = function (config, settings) {
 
   return `${inputsUrl}?team=${teamId}&app=${appId}${settingsQuery}
   `;
-};
+}
