+++ lib/utils/crypto.js
@@ -1,10 +1,13 @@
-const { P256 } = require("../curves");
-const { arrayBufferToBase64 } = require("./datatypes");
+import { Buffer } from 'buffer';
+
+import { P256 } from "../curves";
+import { arrayBufferToBase64 } from "./datatypes";
+
 
 // To compress - record the sign of the `y` point
 // then remove the `y` point and encode the recorded
 // sign in the first bit
-const ecPointCompress = (ecdhRawPublicKey) => {
+export function ecPointCompress (ecdhRawPublicKey) {
   const u8full = new Uint8Array(ecdhRawPublicKey);
   const len = u8full.byteLength;
   const u8 = u8full.slice(0, (1 + len) >>> 1); // drop `y`
@@ -12,7 +15,7 @@ const ecPointCompress = (ecdhRawPublicKey) => {
   return u8;
 };
 
-const deriveSharedSecret = async (ecdh, publicKey, ephemeralPublicKey) => {
+export async function deriveSharedSecret(ecdh, publicKey, ephemeralPublicKey) {
   // Convert to a P256 `CryptoKey` object
   const importedPublicKey = await window.crypto.subtle.importKey(
     "raw",
@@ -64,7 +67,7 @@ const deriveSharedSecret = async (ecdh, publicKey, ephemeralPublicKey) => {
   return hashDigest;
 };
 
-const buildCageKeyFromSuppliedPublicKey = async (publicKey) => {
+export async function buildCageKeyFromSuppliedPublicKey(publicKey) {
   const compressedKey = ecPointCompress(Buffer.from(publicKey, "base64"));
   const compressedKeyString = arrayBufferToBase64(compressedKey);
   return {
@@ -74,8 +77,3 @@ const buildCageKeyFromSuppliedPublicKey = async (publicKey) => {
   };
 };
 
-module.exports = {
-  deriveSharedSecret,
-  ecPointCompress,
-  buildCageKeyFromSuppliedPublicKey,
-};
