+++ lib/utils/datatypes.js
@@ -1,23 +1,24 @@
-const isArray = (data) => data.constructor.name === "Array";
-const isArrayBuffer = (data) =>
+// TODO: Replace with zod
+export const isArray = (data) => data.constructor.name === "Array";
+export const isArrayBuffer = (data) =>
   typeof ArrayBuffer === "function" &&
   (data instanceof ArrayBuffer ||
     Object.prototype.toString.call(data) === "[object ArrayBuffer]");
-const isObject = (data) => typeof data === "object";
-const isObjectStrict = (data) =>
+export const isObject = (data) => typeof data === "object";
+export const isObjectStrict = (data) =>
   isDefined(data) && isObject(data) && !isArray(data) && !isArrayBuffer(data);
-const isString = (data) => typeof data === "string";
-const isEmptyString = (data) => isString(data) && data.length === 0;
-const isDefined = (data) => typeof data !== "undefined" && data !== null;
-const isUndefined = (data) => typeof data === "undefined";
-const isNumber = (data) => typeof data === "number";
-const isBoolean = (data) => typeof data === "boolean";
-const isFile = (data) => data instanceof File || data instanceof Blob;
+export const isString = (data) => typeof data === "string";
+export const isEmptyString = (data) => isString(data) && data.length === 0;
+export const isDefined = (data) => typeof data !== "undefined" && data !== null;
+export const isUndefined = (data) => typeof data === "undefined";
+export const isNumber = (data) => typeof data === "number";
+export const isBoolean = (data) => typeof data === "boolean";
+export const isFile = (data) => data instanceof File || data instanceof Blob;
 
-const isEncryptable = (data) =>
+export const isEncryptable = (data) =>
   isDefined(data) && (isString(data) || isNumber(data) || isBoolean(data));
 
-const getHeaderType = (data) => {
+export const getHeaderType = (data) => {
   if (!isDefined(data)) return;
   if (isArray(data)) return data.constructor.name;
   else {
@@ -25,7 +26,7 @@ const getHeaderType = (data) => {
   }
 };
 
-const ensureString = (data) => {
+export const ensureString = (data) => {
   if (isUndefined(data)) return;
 
   if (!isDefined(data)) return JSON.stringify(data);
@@ -37,7 +38,7 @@ const ensureString = (data) => {
   return JSON.stringify(data);
 };
 
-const utf8ToBase64URL = (str) => {
+export const utf8ToBase64URL = (str) => {
   const base64Encoded = window.btoa(str);
   return base64Encoded
     .replace(/\+/g, "-")
@@ -45,7 +46,7 @@ const utf8ToBase64URL = (str) => {
     .replace(/=/g, "");
 };
 
-const arrayBufferToBase64 = (buffer) => {
+export const arrayBufferToBase64 = (buffer) => {
   var binary = "";
   var bytes = new Uint8Array(buffer);
   var len = bytes.byteLength;
@@ -55,7 +56,7 @@ const arrayBufferToBase64 = (buffer) => {
   return window.btoa(binary);
 };
 
-const base64ToArrayBuffer = (base64) => {
+export const base64ToArrayBuffer = (base64) => {
   var binary_string = window.atob(base64);
   var len = binary_string.length;
   var bytes = new Uint8Array(len);
@@ -63,22 +64,4 @@ const base64ToArrayBuffer = (base64) => {
     bytes[i] = binary_string.charCodeAt(i);
   }
   return bytes.buffer;
-};
-
-module.exports = {
-  isArray,
-  isObject,
-  isObjectStrict,
-  isArrayBuffer,
-  isString,
-  isEmptyString,
-  isDefined,
-  isUndefined,
-  isEncryptable,
-  isFile,
-  getHeaderType,
-  ensureString,
-  utf8ToBase64URL,
-  arrayBufferToBase64,
-  base64ToArrayBuffer,
 };
