+++ lib/utils/domain.js
@@ -1,4 +1,4 @@
-const extractDomainFromUrl = (url) => {
+export default function extractDomainFromUrl(url) {
   var result;
   var match;
   url = url.trim();
@@ -14,5 +14,3 @@ const extractDomainFromUrl = (url) => {
   }
   return result;
 };
-
-module.exports = extractDomainFromUrl;
