+++ lib/utils/errors.js
@@ -1,25 +1,25 @@
-class EvervaultError extends Error {
+export class EvervaultError extends Error {
   constructor(message) {
     super(message);
     this.type = this.constructor.name;
   }
 }
 
-class InitializationError extends EvervaultError {}
+export class InitializationError extends EvervaultError {}
 
-class CryptoError extends EvervaultError {}
+export class CryptoError extends EvervaultError {}
 
-class AccountError extends EvervaultError {}
+export class AccountError extends EvervaultError {}
 
-class ApiKeyError extends EvervaultError {}
+export class ApiKeyError extends EvervaultError {}
 
-class CageKeyError extends EvervaultError {}
+export class CageKeyError extends EvervaultError {}
 
-class RequestError extends EvervaultError {}
+export class RequestError extends EvervaultError {}
 
-class ExceededMaxFileSizeError extends EvervaultError {}
+export class ExceededMaxFileSizeError extends EvervaultError {}
 
-const mapApiResponseToError = ({ statusCode, body }) => {
+export function mapApiResponseToError ({ statusCode, body }) {
   if (statusCode === 401) return new ApiKeyError("Invalid Api Key provided.");
   if (statusCode === 423)
     return new AccountError(
@@ -32,16 +32,4 @@ const mapApiResponseToError = ({ statusCode, body }) => {
         "An error occurred during account creation. Please contact evervault support."
     );
   return new RequestError(`Request returned with status [${statusCode}]`);
-};
-
-module.exports = {
-  EvervaultError,
-  CageKeyError,
-  CryptoError,
-  ApiKeyError,
-  AccountError,
-  InitializationError,
-  mapApiResponseToError,
-  RequestError,
-  ExceededMaxFileSizeError,
 };
