+++ lib/utils/index.js
@@ -1,7 +1,9 @@
-module.exports = {
-  Datatypes: require("./datatypes"),
-  errors: require("./errors"),
-  uuid: require("./uuid"),
-  extractDomain: require("./domain"),
-  cryptoUtils: require("./crypto"),
-};
+export * as Datatypes from "./datatypes";
+export * as errors from "./errors";
+export * as cryptoUtils from "./crypto";
+
+export {default as uuid} from "./uuid";
+export {default as extractDomain} from "./domain";
+export {default as calculateHeight} from "./calculateHeight";
+export {default as constructSource} from "./constructSource";
+
