--- lib/v1/core/index.js
@@ -1,6 +1,0 @@
-module.exports = {
-  Crypto: require("./crypto"),
-  Http: require("./http"),
-  Forms: require("./forms"),
-  Input: require("./inputs"),
-};
