+++ package.json
@@ -1,26 +1,32 @@
 {
-  "name": "evervault-browser-sdk",
-  "version": "1.0.0",
-  "description": "",
-  "main": "./lib/v1/index.js",
+  "name": "@evervault/browser",
+  "version": "2.3.0",
   "engines": {
     "node": ">=16 <=19",
     "pnpm": "~7"
   },
-  "directories": {
-    "lib": "lib"
+  "files": [
+    "dist"
+  ],
+  "main": "dist/index.es.mjs",
+  "module": "dist/index.js",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "require": "./dist/index.umd.cjs"
+    }
   },
   "scripts": {
     "preinstall": "npx only-allow pnpm",
     "test": "mocha",
     "lint": "prettier --check \"./**/*.js\"",
-    "build": "./scripts/build.sh",
-    "dev": "pnpm run build && php -S localhost:5000",
+    "build": "vite build",
+    "dev": "vite",
     "test-serve": "pnpm run test-serve:build && serve -p 3000 ./e2e-app",
-    "test-serve:build": "pnpm run build && cp -r ./dist/v2/index.js ./e2e-app/evervault.bundle.js"
+    "test-serve:build": "pnpm run build && cp -r ./dist/index.js ./e2e-app/evervault.bundle.js"
   },
   "author": "",
-  "license": "ISC",
+  "license": "Apache-2.0",
   "devDependencies": {
     "@playwright/test": "^1.31.2",
     "dotenv": "^16.0.3",
@@ -31,6 +37,11 @@
     "web-file-polyfill": "^1.0.4",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.12",
-    "serve": "^14.2.0"
+    "serve": "^14.2.0",
+    "vite": "^4.2.1"
+  },
+  "dependencies": {
+    "buffer": "^6.0.3",
+    "events": "^3.3.0"
   }
 }
