--- scripts/build.sh
@@ -1,3 +1,0 @@
-find ./lib -maxdepth 1 -type d -exec ./node_modules/.bin/webpack --mode=production --config lib/v1/webpack.config.js   \;
-
-find ./lib -maxdepth 1 -type d -exec ./node_modules/.bin/webpack --mode=production --config lib/v2/webpack.config.js   \;
