+++ test/encryption.js
@@ -1,6 +1,6 @@
 const assert = require("assert");
 require("dotenv").config();
-const Evervault = require("../lib/v2/index.js");
+const Evervault = require("../lib/index.js");
 
 require("./utils").runBrowserJsPolyfills();
