--- test/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Evervault Test</title>
-  </head>
-  <body>
-    evervault Test
-    <script src="../dist/v2/index.js"></script>
-    <script>
-      const ev = new window.Evervault("eab471069135");
-      ev.auto();
-    </script>
-  </body>
-</html>
